+++ changelog.md
@@ -1,6 +1,9 @@
 ==== 1.2.1 ====
 
   * ADDED: Colored Lights support for the Colorful Lamp!
+  * ADDED: More Railcraft Integration
+    - Digital Signal Receiver Box: Same as the Receiver Box driver, but also fires an "aspect_changed" event whenever the box aspect changes
+    - Digital Detector: Fires a "minecart" event whenever a minecart passes
   * CHANGED: setLampColor(0) now turns off the lamp.
 
 ==== 1.2.0 ====